[REFACTOR/#323] 홈화면의 UI 구조를 개편하여 컴포넌트 계층 구조를 간소화합니다.
