--- app/src/main/java/com/sopt/clody/presentation/ui/home/calendar/component/HorizontalDivider.kt
@@ -1,26 +1,0 @@
-package com.sopt.clody.presentation.ui.home.calendar.component
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.unit.Dp
-import androidx.compose.ui.unit.dp
-import com.sopt.clody.ui.theme.ClodyTheme
-
-@Composable
-fun HorizontalDivider(
-    color: Color = ClodyTheme.colors.gray08,
-    thickness: Dp = 6.dp,
-    modifier: Modifier = Modifier,
-) {
-    Box(
-        modifier = modifier
-            .fillMaxWidth()
-            .height(thickness)
-            .background(color),
-    )
-}
