+++ app/src/main/java/com/sopt/clody/presentation/ui/home/component/CalendarDate.kt
@@ -1,0 +1,7 @@
+package com.sopt.clody.presentation.ui.home.component
+
+data class CalendarDate(
+    val date: Int,
+    val month: Int,
+    val year: Int,
+)
