+++ app/src/main/java/com/sopt/clody/presentation/ui/home/screen/HomeViewModel.kt
@@ -14,7 +14,7 @@ import com.sopt.clody.domain.repository.DiaryRepository
 import com.sopt.clody.domain.repository.DraftRepository
 import com.sopt.clody.domain.repository.NotificationRepository
 import com.sopt.clody.domain.repository.ReviewRepository
-import com.sopt.clody.presentation.ui.home.calendar.model.DiaryDateData
+import com.sopt.clody.presentation.ui.home.component.DiaryDateData
 import com.sopt.clody.presentation.ui.setting.notificationsetting.screen.NotificationChangeState
 import com.sopt.clody.presentation.utils.network.ErrorMessageProvider
 import dagger.hilt.android.lifecycle.HiltViewModel
